requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.25
numpy==1.20.2
pandas==1.2.4
protobuf==3.17.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.1.0
Werkzeug==1.0.1
